Exploit APK initial (#182)
* initial exploit_apk

* add exploit apk

* Working empty exploit APK

* add app folder

* Refactor helpers out to specific modules

* Adding in missing helpers

* File permission written and tested (#168)

* add tests for file permissions

* add plugin for file permissions

* add tests for file permissions

* update helpers

* fix spacing from merge conflict

* Pending intent written and tested (#166)

* add scanner checks for intent

* add implicit intent checks

* add implicit intent tests

* try/catch file open and change unused variable to _

* remove broadcast check for apk due to decompilation

* remove unnecessary code to make more compact

* add scanner checks for intent

* add implicit intent checks

* add implicit intent tests

* try/catch file open and change unused variable to _

* address pr comments, compile regex and use _

* remove merge artifact

* update plugins to use helper, remove file locations from description, exit checks if manifest doesnt exist, catch another exception

* initial exploit_apk

* change scanner to use categories instead of functions

* merge

* update helpers

* working exploit code

* fix decompiler tests

* fix test_scanner

* fix decompiler

* change get_manifest

* remove build directory

* remove build directory

* remove build directory

* comment apk_builder tests because SDK isn't available

+++ tests/test_decompiler/test_decompiler.py
@@ -4,6 +4,7 @@ import shutil
 import pytest
 
 from qark.decompiler import decompiler
+from qark.decompiler.decompiler import Decompiler
 from qark.decompiler.external_decompiler import JDCore, CFR, Procyon
 
 
@@ -67,28 +68,39 @@ def test_get_java_version():
     assert not java_version.endswith('"')
 
 
-def test_ran_unpack_apk(decompiler):
+def test_unpack_apk(decompiler):
     classes_dex_path = os.path.join(decompiler.build_directory, "classes.dex")
-    assert classes_dex_path == decompiler.dex_path
+    assert classes_dex_path == decompiler._unpack_apk()
     assert os.path.isfile(classes_dex_path)
 
     shutil.rmtree(decompiler.build_directory)
     assert not os.path.isdir(decompiler.build_directory)
 
+    with pytest.raises(ValueError):
+        bad_decompiler = Decompiler("1")
+        bad_decompiler._unpack_apk()
 
-def test_ran_apktool(decompiler):
+
+def test_run_apktool(decompiler):
+    if os.path.exists(decompiler.build_directory):
+        shutil.rmtree(decompiler.build_directory)
+
+    decompiler.run_apktool()
+    assert os.path.isdir(decompiler.build_directory)
     assert os.path.isfile(os.path.join(decompiler.build_directory, "AndroidManifest.xml"))
 
     shutil.rmtree(decompiler.build_directory)
     assert not os.path.isdir(decompiler.build_directory)
 
 
-def test_ran_dex2jar(decompiler):
-    assert os.path.isfile(decompiler.jar_path)
-    assert decompiler.jar_path.endswith(".jar")
+def test_run_dex2jar(decompiler):
+    if os.path.exists(decompiler.build_directory):
+        shutil.rmtree(decompiler.build_directory)
 
-    shutil.rmtree(decompiler.build_directory)
-    assert not os.path.isdir(decompiler.build_directory)
+    decompiler.dex_path = decompiler._unpack_apk()
+    jar_file = decompiler._run_dex2jar()
+    assert os.path.isfile(jar_file)
+    assert jar_file.endswith(".jar")
 
 
 def test_decompile(decompiler):
